# Improve dependency management and archive generation