--- a/elyra/pipeline/processor_kfp.py
+++ b/elyra/pipeline/processor_kfp.py
@@ -269,10 +269,12 @@ class KfpPipelineProcessor(PipelineProcessor):
         archive_artifact_name = self._get_dependency_archive_name(operation)
         archive_source_dir = self._get_dependency_source_dir(operation)
 
+        dependencies = [os.path.basename(operation.filename)]
+        dependencies.extend(operation.dependencies)
+
         archive_artifact = create_temp_archive(archive_name=archive_artifact_name,
                                                source_dir=archive_source_dir,
-                                               primary_file=os.path.basename(operation.filename),
-                                               dependencies=operation.dependencies,
+                                               filenames=dependencies,
                                                recursive=operation.include_subdirectories)
 
         return archive_artifact